# docs/openapi/v1/api.yaml
openapi: 3.0.3
info:
  title: Shop API
  description: |
    REST API магазина бытовой техники.

    API управляет клиентами, товарами, поставщиками и изображениями товаров.
    Идентификаторы ресурсов имеют формат UUID, даты — ISO 8601. Ошибки
    возвращаются как JSON-объекты с текстом сообщения.
  version: 1.0.0
servers:
- url: http://localhost:{port}/api/v1
  description: Локальный сервер
  variables:
    port:
      default: '8080'
      description: Порт HTTP-сервера
...
paths:
...
  "/products":
    post:
      tags:
      - Products
      summary: Добавить товар
      operationId: createProduct
      parameters:
      - "$ref": "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductCreate"
            example:
              name: Холодильник Frost 3000
              category: Холодильники
              price: 54990.0
              available_stock: 12
              last_update_date: '2026-08-26'
              supplier_id: 7116b493-97ab-472f-8ae2-b34ed382619c
      responses:
        '201':
          description: Товар создан
          headers:
            Location:
              description: URL созданного товара
              schema:
                type: string
                format: uri
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Product"
        '400':
          "$ref": "#/components/responses/ValidationError"
        '404':
          description: Поставщик не найден
          content:
            application/json:
              schema:
                "$ref": "#/components/schemas/Error"
              example:
                code: SUPPLIER_NOT_FOUND
                message: supplier not found
        '409':
          "$ref": "#/components/responses/IdempotencyConflict"
        '500':
          "$ref": "#/components/responses/InternalError"
    get:
      tags:
      - Products
      summary: Получить все доступные товары
      description: |
        Возвращает товары с `available_stock` больше нуля. Если таких товаров
        нет, возвращается пустой массив.
      operationId: getAvailableProducts
      responses:
        '200':
          description: Список доступных товаров
          content:
            application/json:
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/Product"
              examples:
                result:
                  summary: Доступные товары
                  value:
                  - id: '0866613a-d3d6-449d-99d4-d821c5755b0b'
                    name: Холодильник Frost 3000
                    category: Холодильники
                    price: 54990.0
                    available_stock: 12
                    last_update_date: '2026-08-26'
                    supplier_id: 7116b493-97ab-472f-8ae2-b34ed382619c
                    image_id: 1e07a39f-d7d1-49f0-bc33-80f53f6857d6
                empty:
                  summary: Доступные товары отсутствуют
                  value: []
        '500':
          "$ref": "#/components/responses/InternalError"
...
components:
  schemas:
...
    Error:
      type: object
      required:
      - code
      - message
      properties:
        code:
          type: string
          description: Стабильный машиночитаемый код ошибки
          example: RESOURCE_NOT_FOUND
        message:
          type: string
          description: Понятное человеку описание ошибки
...
    Product:
      type: object
      required:
      - id
      - name
      - category
      - price
      - available_stock
      - last_update_date
      - supplier_id
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          minLength: 1
          maxLength: 255
        category:
          type: string
          minLength: 1
          maxLength: 100
        price:
          type: number
          format: double
          minimum: 0
        available_stock:
          type: integer
          format: int64
          minimum: 0
        last_update_date:
          type: string
          format: date
        supplier_id:
          type: string
          format: uuid
        image_id:
          type: string
          format: uuid
          nullable: true
          readOnly: true
      example:
        id: '0866613a-d3d6-449d-99d4-d821c5755b0b'
        name: Холодильник Frost 3000
        category: Холодильники
        price: 54990.0
        available_stock: 12
        last_update_date: '2026-08-26'
        supplier_id: 7116b493-97ab-472f-8ae2-b34ed382619c
        image_id: 1e07a39f-d7d1-49f0-bc33-80f53f6857d6
    ProductCreate:
      type: object
      additionalProperties: false
      required:
      - name
      - category
      - price
      - available_stock
      - supplier_id
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        category:
          type: string
          minLength: 1
          maxLength: 100
        price:
          type: number
          format: double
          minimum: 0
        available_stock:
          type: integer
          format: int64
          minimum: 0
        last_update_date:
          type: string
          format: date
          description: Если не задана, сервер использует текущую дату
        supplier_id:
          type: string
          format: uuid
...
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      description: |
        Уникальный ключ операции. Повтор с тем же ключом и телом возвращает
        сохранённый результат; тот же ключ с другим телом приводит к `409`.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 255
      example: 550e8400-e29b-41d4-a716-446655440000
...
  responses:
    IdempotencyConflict:
      description: Ключ идемпотентности уже использован для другого запроса
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            code: IDEMPOTENCY_KEY_REUSED
            message: idempotency key is already used with a different request
    ValidationError:
      description: Тело запроса не прошло валидацию
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            code: VALIDATION_ERROR
            message: validation failed
...
    InternalError:
      description: Внутренняя ошибка сервера
      content:
        application/json:
          schema:
            "$ref": "#/components/schemas/Error"
          example:
            code: INTERNAL_ERROR
            message: internal server error
...